Separate recoverable from unrecoverable errors
This start splitting the handling of recoverable and unrecoverable errors.
Recoverable errors are now accumulated on the parser and interpret objects and
can be retrieved after processing has completed. Unrecoverable errors are
returned immediately through result types. This opens up the door for the
parser synchronizing and continuing to process after encountering a recoverable
error, although this is not yet used. This may also be useful in the
interpreter for treating return statements as a type of result.

The parse and interpret entry points can now yield multiple errors. These are
handled awkwardly in this change.