Separate AST lifetime from source, start on lox functions
This separates the lifetime of AST elements from the original source by
introducing string copies for names instead of referencing the source buffer.
This allows keeping an interpreter with values around longer than source
and should make it simpler to define a function with retained statements.

This also partially implements callable functions defined in Lox.