Allocate and remember unique identifiers for expressions
To implement and use scope resolution we need a way to identify expressions
during the resolution AST pass and then later identify the same expressions
during evaluation. This adds an 'id' field which is stored in each expression
and can be accessed (but not set) outside the 'ast' crate.

Expressions are stored in the heap (via Box or Rc) and so another approach would
be to use the address of each expression as its identifier. This would work for
the resolution + evaluation pass as we do not allocate or deallocate any new
expressions in between resolution or evaluation. by_address gets part of the
way, but I'm not sure how to annotate the lifetimes correctly. We would need to
convince the compiler that the addresses of expressions live longer than the use
of addresses in the interpreter's resolution map.